# build/resources/main/static/openapi-spec.yaml
openapi: 3.0.3
info:
  title: 캠퍼스 하우스 API
  description: 대학생 자취 커뮤니티 앱의 백엔드 API 문서
  version: 1.5.0
  contact:
    name: Campus House Team
    email: contact@campushouse.com
    url: https://campushouse.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: 개발 서버
  - url: https://api.campushouse.com
    description: 프로덕션 서버
security:
  - bearerAuth: []
paths:
...
  /api/boards/{type}/posts:
    post:
      tags:
        - 게시글
      summary: 게시글 작성
      description: 새로운 게시글을 작성합니다. APARTMENT와 QUESTION 게시판은 거주지 인증이 필요합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          description: 게시판 타입
          schema:
            type: string
            enum: [APARTMENT, QUESTION, LOCAL, TRANSFER]
            example: APARTMENT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: 게시글 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: 잘못된 요청
        '401':
          description: 인증 필요
        '403':
          description: 거주지 인증 필요

    get:
      tags:
        - 게시글
      summary: 게시글 목록 조회
      description: 특정 게시판의 게시글 목록을 조회합니다. APARTMENT와 QUESTION 게시판은 거주지 인증이 필요합니다.
      parameters:
        - name: type
          in: path
          required: true
          description: 게시판 타입
          schema:
            type: string
            enum: [APARTMENT, QUESTION, LOCAL, TRANSFER]
            example: APARTMENT
        - name: page
          in: query
          description: 페이지 번호
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: 페이지당 게시글 수
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          description: 정렬 기준
          schema:
            type: string
            default: "createdAt,desc"
      responses:
        '200':
          description: 게시글 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
  /api/boards/{type}/posts/{postId}:
    get:
      tags:
        - 게시글
      summary: 게시글 상세 조회
      description: 특정 게시글의 상세 정보를 조회합니다.
      parameters:
        - name: type
          in: path
          required: true
          description: 게시판 타입
          schema:
            type: string
            enum: [APARTMENT, QUESTION, LOCAL, TRANSFER]
            example: APARTMENT
        - name: postId
          in: path
          required: true
          description: 게시글 ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 게시글 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '404':
          description: 게시글을 찾을 수 없음

    put:
      tags:
        - 게시글
      summary: 게시글 수정
      description: 게시글을 수정합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          description: 게시판 타입
          schema:
            type: string
            enum: [APARTMENT, QUESTION, LOCAL, TRANSFER]
            example: APARTMENT
        - name: postId
          in: path
          required: true
          description: 게시글 ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePostRequest'
      responses:
        '200':
          description: 게시글 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: 잘못된 요청
        '401':
          description: 인증 필요
        '403':
          description: 권한 없음
        '404':
          description: 게시글을 찾을 수 없음

    delete:
      tags:
        - 게시글
      summary: 게시글 삭제
      description: 게시글을 삭제합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          description: 게시판 타입
          schema:
            type: string
            enum: [APARTMENT, QUESTION, LOCAL, TRANSFER]
            example: APARTMENT
        - name: postId
          in: path
          required: true
          description: 게시글 ID
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 게시글 삭제 성공
        '401':
          description: 인증 필요
        '403':
          description: 권한 없음
        '404':
          description: 게시글을 찾을 수 없음
...
  /api/buildings/{buildingId}/questions:
    post:
      tags:
        - 건물 정보
      summary: 건물별 질문 작성
      description: 특정 건물에 대한 질문을 작성합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: buildingId
          in: path
          required: true
          description: 건물 ID
          schema:
            type: integer
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: 질문 제목
                  example: "이 건물 주차장은 어떻게 되나요?"
                content:
                  type: string
                  description: 질문 내용
                  example: "주차장이 있는지, 월 주차비는 얼마인지 궁금합니다."
      responses:
        '200':
          description: 질문 작성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '400':
          description: 잘못된 요청
        '401':
          description: 인증 필요
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT 토큰을 입력하세요. 예: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
  schemas:
...
    User:
      type: object
      description: 사용자 정보
      properties:
        id:
          type: integer
          description: 사용자 ID
          example: 1
        email:
          type: string
          format: email
          description: 이메일
          example: user@example.com
        nickname:
          type: string
          description: 닉네임
          example: 홍길동
        userType:
          type: string
          enum: [STUDENT, GRADUATE, OTHER]
          description: 사용자 타입
          example: STUDENT
        location:
          type: string
          description: 거주 지역
          example: 서울시강남구
        university:
          type: string
          description: 대학교
          example: 경희대학교
        major:
          type: string
          description: 전공
          example: 컴퓨터공학과
        profileImage:
          type: string
          description: 프로필 이미지 URL
          example: https://example.com/profile.jpg
        createdAt:
          type: string
          format: date-time
          description: 가입일시
          example: "2024-01-01T00:00:00"
        updatedAt:
          type: string
          format: date-time
          description: 수정일시
          example: "2024-01-01T00:00:00"
    CreatePostRequest:
      type: object
      description: 게시글 작성 요청
      required:
        - title
        - content
      properties:
        title:
          type: string
          description: 게시글 제목
          example: "새로운 게시글입니다"
        content:
          type: string
          description: 게시글 내용
          example: "게시글 내용입니다"
        imageUrl:
          type: string
          description: 이미지 URL
          example: "https://example.com/image.jpg"
    UpdatePostRequest:
      type: object
      description: 게시글 수정 요청
      properties:
        title:
          type: string
          description: 게시글 제목
          example: "수정된 게시글 제목"
        content:
          type: string
          description: 게시글 내용
          example: "수정된 게시글 내용"
        imageUrl:
          type: string
          description: 이미지 URL
          example: "https://example.com/updated-image.jpg"
    Post:
      type: object
      description: 게시글 정보
      properties:
        id:
          type: integer
          description: 게시글 ID
          example: 1
        title:
          type: string
          description: 게시글 제목
          example: "새로운 게시글입니다"
        content:
          type: string
          description: 게시글 내용
          example: "게시글 내용입니다"
        imageUrl:
          type: string
          description: 이미지 URL
          example: "https://example.com/image.jpg"
        author:
          $ref: '#/components/schemas/User'
        building:
          $ref: '#/components/schemas/Building'
          description: 건물 정보 (QUESTION, TRANSFER 게시판에서만 사용)
        boardType:
          type: string
          enum: [APARTMENT, QUESTION, LOCAL, TRANSFER]
          description: 게시판 타입
          example: APARTMENT
        viewCount:
          type: integer
          description: 조회수
          example: 100
        likeCount:
          type: integer
          description: 좋아요 수
          example: 15
        commentCount:
          type: integer
          description: 댓글 수
          example: 5
        isNew:
          type: boolean
          description: 새 게시글 여부
          example: true
        createdAt:
          type: string
          format: date-time
          description: 작성일시
          example: "2024-01-01T00:00:00"
        updatedAt:
          type: string
          format: date-time
          description: 수정일시
          example: "2024-01-01T00:00:00"
...
    Building:
      type: object
      description: 건물 정보
      properties:
        id:
          type: integer
          description: 건물 ID
          example: 1
        buildingName:
          type: string
          description: 건물명
          example: "캠퍼스 하우스 A동"
        address:
          type: string
          description: 주소
          example: "서울시 강남구 테헤란로 123"
        detailAddress:
          type: string
          description: 상세주소
          example: "101호"
        latitude:
          type: number
          format: double
          description: 위도
          example: 37.5665
        longitude:
          type: number
          format: double
          description: 경도
          example: 126.9780
        deposit:
          type: number
          format: decimal
          description: 보증금 (만원)
          example: 1000
        monthlyRent:
          type: number
          format: decimal
          description: 월세 (만원)
          example: 50
        jeonse:
          type: number
          format: decimal
          description: 전세 (만원)
          example: 10000
        managementFee:
          type: number
          format: decimal
          description: 관리비 (만원)
          example: 5
        households:
          type: integer
          description: 세대수
          example: 200
        heatingType:
          type: string
          description: 난방방식
          example: "개별난방"
        parkingSpaces:
          type: integer
          description: 주차대수
          example: 50
        elevators:
          type: integer
          description: 승강기대수
          example: 2
        buildingUsage:
          type: string
          description: 건축물용도
          example: "오피스텔 84%, 기타제2종근린생활시설 16%"
        approvalDate:
          type: string
          format: date-time
          description: 사용승인일
          example: "2020-01-15T00:00:00"
        completionDate:
          type: string
          format: date-time
          description: 준공일
          example: "2019-12-01T00:00:00"
        nearbyConvenienceStores:
          type: integer
          description: 반경 1km 이내 편의점 개수 (외부 API 연동 예정)
          example: 5
        nearbyMarts:
          type: integer
          description: 반경 1km 이내 마트 개수 (외부 API 연동 예정)
          example: 2
        nearbyHospitals:
          type: integer
          description: 반경 1km 이내 병원 개수 (외부 API 연동 예정)
          example: 1
        schoolWalkingTime:
          type: integer
          description: 학교까지 걸리는 시간 (분)
          example: 15
        stationWalkingTime:
          type: integer
          description: 영통역까지 걸리는 시간 (분)
          example: 8
        scrapCount:
          type: integer
          description: 스크랩수
          example: 15
        createdAt:
          type: string
          format: date-time
          description: 생성일시
          example: "2024-01-01T00:00:00"
        updatedAt:
          type: string
          format: date-time
          description: 수정일시
          example: "2024-01-01T00:00:00"
...